requirements.txt
opencv-python==4.12.0.88
pillow==12.0.0
numpy==2.2.6
tqdm==4.67.1
ultralytics
PyQt5
matplotlib
reportlab
flask
flask-cors
requests
pyyaml
